+++ D6831-LN01-InputOutput-JKT/Problem R: The Registration Begin/answer.c
@@ -2,13 +2,13 @@
 
 int main() {
     char N[11], M[101], C;
-    int O;
+    unsigned short O;
     scanf("%10s\n%100[^\n]", N, M);
     getchar();
-    scanf("%c %d", &C, &O);
+    scanf("%c %hu", &C, &O);
     printf("Id    : %s\n", N);
     printf("Name  : %s\n", M);
     printf("Class : %c\n", C);
-    printf("Num   : %d\n", O);
+    printf("Num   : %hu\n", O);
     return 0;
 }
